Fix statvfs cross-platform: cast fields to u64 for macOS compatibility
On macOS, statvfs fields (f_blocks, f_bavail, f_bfree, f_files, f_ffree, f_frsize)
are u32, while on Linux they are u64. Explicit `as u64` casts work on both platforms.

+++ crates/shoal-store/src/file_store.rs
@@ -194,15 +194,15 @@ fn statvfs(path: &Path) -> Result<StorageCapacity, StoreError> {
             return Err(StoreError::Io(std::io::Error::last_os_error()));
         }
 
-        let block_size = stat.f_frsize;
-        let total = stat.f_blocks * block_size;
-        let available = stat.f_bavail * block_size;
+        let block_size = stat.f_frsize as u64;
+        let total = stat.f_blocks as u64 * block_size;
+        let available = stat.f_bavail as u64 * block_size;
         // f_bfree includes blocks reserved for root; f_bavail is what unprivileged users can use.
-        let free = stat.f_bfree * block_size;
+        let free = stat.f_bfree as u64 * block_size;
         let used = total.saturating_sub(free);
 
-        let inodes_total = stat.f_files;
-        let inodes_free = stat.f_ffree;
+        let inodes_total = stat.f_files as u64;
+        let inodes_free = stat.f_ffree as u64;
 
         // Warn when inode usage exceeds 80%.
         if inodes_total > 0 {
